# docs/swagger.yaml
basePath: /v1
definitions:
...
  benchmarkController.RunResponse:
    description: Result of a benchmark submission phase.
    properties:
      concurrency:
        type: integer
      config_fingerprint:
        type: string
      export_url:
        type: string
      failed_jobs:
        type: integer
      rate_per_second:
        type: number
      requested_jobs:
        type: integer
      run_id:
        type: string
      submit_wall_clock_ns:
        type: integer
      submitted_jobs:
        type: integer
      target_host:
        type: string
      variant:
        type: string
      workload_id:
        type: string
    type: object
...
  payload.RESTPayload:
    properties:
      callback_url:
        description: Optional per-job destination for forwarding aggregated logs/results.
          Must be an absolute http/https URL with a host.
        format: uri
        type: string
      id:
        description: Unique job identifier
        type: string
      metadata:
        additionalProperties:
          type: string
        description: Additional job-level metadata
        type: object
      name:
        description: Human-readable job name
        type: string
      priority:
        description: Priority level (1=low, 2=medium, 3+=high)
        type: integer
      steps:
        description: Ordered list of steps to execute
        items:
          $ref: '#/definitions/payload.Step'
        type: array
      timestamp:
        description: Job creation timestamp
        type: string
    required:
    - name
    type: object
  payload.Step:
    properties:
      continue_on_error:
        description: Whether to continue with the next step if this step fails
        type: boolean
      cpu_limit:
        description: CPU limit in millicores (e.g., 1000 = 1 CPU core)
        type: integer
      id:
        description: Step execution order (starts at 1)
        type: integer
      image:
        description: Container image to use (e.g., "alpine:latest")
        type: string
      memory_limit:
        description: Memory limit (e.g., "512M", "2G")
        type: string
      metadata:
        additionalProperties:
          type: string
        description: Step-specific environment variables and metadata
        type: object
      name:
        description: Human-readable step name
        type: string
      script:
        description: Shell script to execute in the container
        type: string
    type: object
  persister.Run:
    properties:
      benchmarker_version:
        type: string
      commit_hash:
        type: string
      concurrency:
        type: integer
      config_fingerprint:
        type: string
      failed_jobs:
        type: integer
      finished_at_ns:
        type: integer
      notes:
        type: string
      priority:
        type: integer
      rate_per_second:
        type: number
      requested_jobs:
        type: integer
      run_id:
        type: string
      started_at_ns:
        type: integer
      submitted_jobs:
        type: integer
      target_host:
        type: string
      variant:
        type: string
      workload_id:
        type: string
    type: object
  response.ErrorMessage:
    properties:
      error:
        example: Failed to parse UUID
        type: string
    type: object
...
  response.ServerErrorMessage:
    properties:
      error:
        example: Internal server error
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: yu.shuaiwei@tum.de
    name: Shuaiwei Yu
    url: https://github.com/Mtze
  description: 'Measurement instrument for comparing CI job-execution variants. Jobs
    are submitted over REST and completions are observed over REST: the system under
    test posts a terminal status callback to /v1/callback and the benchmarker stamps
    arrival on its own clock. Analysis is done on the raw rows from /v1/export/jobs,
    never on the aggregate endpoints.'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/Hades-Scheduler/CI-Benchmarker
  title: CI-Benchmarker API
  version: "2.0"
paths:
...
  /benchmark/hades-docker:
    post:
      consumes:
      - application/json
      description: Submits jobs to a Hades instance running the Docker executor.
      parameters:
      - description: Hades REST endpoint to submit to
        in: query
        name: host
        required: true
        type: string
      - default: 1
        description: Number of jobs to submit
        in: query
        name: count
        type: integer
      - description: Run identifier; generated when omitted
        in: query
        name: run_id
        type: string
      - description: Maximum in-flight submissions
        in: query
        name: concurrency
        type: integer
      - description: Offered submissions per second
        in: query
        name: rate
        type: number
      - description: Identifier for the submitted payload
        in: query
        name: workload_id
        type: string
      - description: Optional run tag
        in: query
        name: commit_hash
        type: string
      - description: Where Hades should POST terminal status; defaults to CALLBACK_BASE_URL
          + /v1/callback
        in: query
        name: status_callback_url
        type: string
      - description: Job payload to submit
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/payload.RESTPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/benchmarkController.RunResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ServerErrorMessage'
      summary: Benchmark Hades with the Docker executor
      tags:
      - benchmark
  /benchmark/hades-k8s:
    post:
      consumes:
      - application/json
      description: Submits jobs to a Hades instance running the Kubernetes executor.
      parameters:
      - description: Hades REST endpoint to submit to
        in: query
        name: host
        required: true
        type: string
      - default: 1
        description: Number of jobs to submit
        in: query
        name: count
        type: integer
      - description: Run identifier; generated when omitted
        in: query
        name: run_id
        type: string
      - description: Maximum in-flight submissions
        in: query
        name: concurrency
        type: integer
      - description: Offered submissions per second
        in: query
        name: rate
        type: number
      - description: Identifier for the submitted payload
        in: query
        name: workload_id
        type: string
      - description: Optional run tag
        in: query
        name: commit_hash
        type: string
      - description: Where Hades should POST terminal status; defaults to CALLBACK_BASE_URL
          + /v1/callback
        in: query
        name: status_callback_url
        type: string
      - description: Job payload to submit
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/payload.RESTPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/benchmarkController.RunResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ServerErrorMessage'
      summary: Benchmark Hades with the Kubernetes executor
      tags:
      - benchmark
...
  /export/callbacks:
    get:
      description: Streams every recorded terminal callback as JSON Lines, including
        ones that never matched a submission. Set only_unmatched=true to list just
        the orphans, which is worth checking after every run.
      parameters:
      - description: Restrict to one run
        in: query
        name: run_id
        type: string
      - default: false
        description: Only callbacks with no matching submission
        in: query
        name: only_unmatched
        type: boolean
      produces:
      - text/plain
      responses:
        "200":
          description: JSONL stream
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ServerErrorMessage'
      summary: Export raw callbacks
      tags:
      - export
  /export/jobs:
    get:
      description: Streams one row per submission attempt, joined to its terminal
        callback, as JSON Lines (default) or CSV. Jobs that never called back are
        included with completed=false; excluding them is what produced survivorship
        bias in the previous design. All analysis and every figure should be generated
        from this endpoint.
      parameters:
      - description: Restrict to one run; omit to export every run
        in: query
        name: run_id
        type: string
      - default: jsonl
        description: jsonl or csv
        enum:
        - jsonl
        - csv
        in: query
        name: format
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: JSONL or CSV stream
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorMessage'
      summary: Export raw per-job records
      tags:
      - export
  /export/runs:
    get:
      description: Returns every recorded run with its provenance, newest first.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/persister.Run'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ServerErrorMessage'
      summary: List benchmark runs
      tags:
      - export
...
schemes:
- http
- https
swagger: "2.0"